Fix to consider the case where there may not be a prefix or aspa pdu between a cache-response and end-of-data.

+++ tools/checks.sh
@@ -270,8 +270,12 @@ check_cache_response() {
 		echo "$i" >> "$EXPECTED"
 	done
 
-	sort "$ACTUAL.tmp" > "$ACTUAL"
-	rm "$ACTUAL.tmp"
+	if [ -f "$ACTUAL.tmp" ]; then
+		sort "$ACTUAL.tmp" > "$ACTUAL"
+		rm "$ACTUAL.tmp"
+	else
+		:> "$ACTUAL"
+	fi
 
 	ck_inc
 	diff "$EXPECTED" "$ACTUAL" > "$DIFF" ||
